Update package version and add steamApiKey to TradeWebsocket constructor
The package version was updated from 1.3.5 to 1.3.6 in the package.json file. In the TradeWebsocket class, a new private property called steamApiKey was added to the constructor arguments and initialized in the constructor function. Additionally, three new JSON objects were sent through the websocket connection with custom source fields and identity_secret/identify_secret properties set to true for two of them.