Updated package and TypeScript configuration
The package.json file has been updated to include the "type" field set as "module". This change allows Node.js to treat the files in this project as ES modules. 

In addition, the tsconfig.json file's compilerOptions target has been corrected from "ESNEXT" to "ESNext". This ensures that TypeScript will compile down to a more recent version of ECMAScript.