Replaced axios with undici for HTTP requests
The HTTP client library has been switched from axios to undici. This change includes updating the constructor to initialize an instance of the undici Client instead of creating a new https.Agent. The post and get methods have also been updated to use this new client, including changes in how the request URL is constructed and how responses are handled.